requirements.txt
Django==5.1.6
djangorestframework==3.15.2
djangorestframework-simplejwt==5.5.0
django-cors-headers==4.7.0
python-dotenv==1.0.1
drf-yasg==1.21.9
PyYAML==6.0.2
pytz==2025.1 